// src/types/FeaturedProject.ts
export type FeaturedProject = {
    id: string;
    title: string;
    tagline: string;
    bannerUrl: string;
    description: string;
    ownerAvatarUrl: string;
    ownerDisplayName: string;
    repoId: number;
    repoLanguage: string;
    repoOpenTickets: number;
    repoStars: number;
    repoUrl: string;
    repoPushedAt: Date;
}